Combine the two instance sign in pages
https://harperdb.atlassian.net/browse/STUDIO-242
